% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodistpar.R
\name{geodistpar}
\alias{geodistpar}
\title{geodistpar}
\usage{
geodistpar(x, y, measure = "cheap")
}
\arguments{
\item{x}{Rectangular object}

\item{y}{Rectangular object}

\item{measure}{One of "haversine" "vincenty", "geodesic", or "cheap"
specifying desired method of geodesic distance calculation}
}
\description{
Generate a distance matrix
}
